This is a large commit. I changed the .Rmd structure, so one .Rmd per analysis.
- Timeline_FolderStructureAndDataCleaning
This file only contains the code for setting up the folder structure and cleaning the data

- Timeline_DescriptiveStatistics
This file only contains the code for descriptive statistics. I also added some questions that the file is dealing with (Important)

- Timeline_ClusterAnalysis
This file only contains the code for the cluster analysis. I have to work more with this in the fufure!